convert packer to lazy.nvim
