+++ dot_config/nvim/init.lua
@@ -1,113 +1,114 @@
--- Install packer
-local install_path = vim.fn.stdpath 'data' .. '/site/pack/packer/start/packer.nvim'
-local is_bootstrap = false
-if vim.fn.empty(vim.fn.glob(install_path)) > 0 then
-  is_bootstrap = true
-  vim.fn.system { 'git', 'clone', '--depth', '1', 'https://github.com/wbthomason/packer.nvim', install_path }
-  vim.cmd [[packadd packer.nvim]]
-end
-require('packer').startup(function(use) -- Package manager
-  use 'wbthomason/packer.nvim'
-
-  use { -- LSP Configuration & Plugins
-    'neovim/nvim-lspconfig',
-    requires = {
-      -- Automatically install LSPs to stdpath for neovim
-      'williamboman/mason.nvim',
-      'williamboman/mason-lspconfig.nvim',
-
-      -- Useful status updates for LSP
-      { 'j-hui/fidget.nvim', tag = 'legacy' },
-
-      -- Additional lua configuration, makes nvim stuff amazing
-      'folke/neodev.nvim',
-    },
-  }
-
-  use { -- Autocompletion
-    'hrsh7th/nvim-cmp',
-    requires = { 'hrsh7th/cmp-nvim-lsp', 'L3MON4D3/LuaSnip', 'saadparwaiz1/cmp_luasnip' },
-  }
-
-  use { -- Highlight, edit, and navigate code
-    'nvim-treesitter/nvim-treesitter',
-    run = function()
-      pcall(require('nvim-treesitter.install').update { with_sync = true })
-    end,
-  }
-
-  use { -- Additional text objects via treesitter
-    'nvim-treesitter/nvim-treesitter-textobjects',
-    after = 'nvim-treesitter',
-  }
-
-  -- Git related plugins
-  use 'tpope/vim-fugitive'
-  use 'tpope/vim-rhubarb'
-  use 'lewis6991/gitsigns.nvim'
-
-  -- Look good
-  use 'nvim-lualine/lualine.nvim' -- Fancier statusline
-  use 'lukas-reineke/indent-blankline.nvim' -- Add indentation guides even on blank lines
-  use 'numToStr/Comment.nvim' -- "gc" to comment visual regions/lines
-  use 'tpope/vim-sleuth' -- Detect tabstop and shiftwidth automatically
-  use { "catppuccin/nvim", as = "catppuccin" }
-
-  -- Go
-  use 'ray-x/go.nvim'
-  use 'ray-x/guihua.lua'
-
-  -- File tree
-  use 'nvim-tree/nvim-web-devicons'
-  use 'nvim-tree/nvim-tree.lua'
-
-  -- Fuzzy Finder (files, lsp, etc)
-  use { 'nvim-telescope/telescope.nvim', branch = '0.1.x', requires = { 'nvim-lua/plenary.nvim' } }
-
-  -- Fuzzy Finder Algorithm which requires local dependencies to be built. Only load if `make` is available
-  use { 'nvim-telescope/telescope-fzf-native.nvim', run = 'make', cond = vim.fn.executable 'make' == 1 }
-
-  use({
-      "aserowy/tmux.nvim",
-      config = function() return require("tmux").setup() end
-  })
-
-  -- Add custom plugins to packer from ~/.config/nvim/lua/custom/plugins.lua
-  local has_plugins, plugins = pcall(require, 'custom.plugins')
-  if has_plugins then
-    plugins(use)
-  end
-
-  if is_bootstrap then
-    require('packer').sync()
-  end
-end)
-
-vim.g.loaded_netrw = 1
-vim.g.loaded_netrwPlugin = 1
-
-
--- When we are bootstrapping a configuration, it doesn't
--- make sense to execute the rest of the init.lua.
+-- -- Install packer
+-- local install_path = vim.fn.stdpath 'data' .. '/site/pack/packer/start/packer.nvim'
+-- local is_bootstrap = false
+-- if vim.fn.empty(vim.fn.glob(install_path)) > 0 then
+--   is_bootstrap = true
+--   vim.fn.system { 'git', 'clone', '--depth', '1', 'https://github.com/wbthomason/packer.nvim', install_path }
+--   vim.cmd [[packadd packer.nvim]]
+-- end
+-- require('packer').startup(function(use) -- Package manager
+--   use 'wbthomason/packer.nvim'
 --
--- You'll need to restart nvim, and then it will work.
-if is_bootstrap then
-  print '=================================='
-  print '    Plugins are being installed'
-  print '    Wait until Packer completes,'
-  print '       then restart nvim'
-  print '=================================='
-  return
-end
-
--- Automatically source and re-compile packer whenever you save this init.lua
-local packer_group = vim.api.nvim_create_augroup('Packer', { clear = true })
-vim.api.nvim_create_autocmd('BufWritePost', {
-  command = 'source <afile> | PackerCompile',
-  group = packer_group,
-  pattern = vim.fn.expand '$MYVIMRC',
-})
-
+--   use { -- LSP Configuration & Plugins - converted (i hope...)
+--     'neovim/nvim-lspconfig',
+--     requires = {
+--       -- Automatically install LSPs to stdpath for neovim
+--       'williamboman/mason.nvim',
+--       'williamboman/mason-lspconfig.nvim',
+--
+--       -- Useful status updates for LSP
+--       { 'j-hui/fidget.nvim', tag = 'legacy' },
+--
+--       -- Additional lua configuration, makes nvim stuff amazing
+--       'folke/neodev.nvim',
+--     },
+--   }
+--
+--   use { -- Autocompletion - converted
+--     'hrsh7th/nvim-cmp',
+--     requires = { 'hrsh7th/cmp-nvim-lsp', 'L3MON4D3/LuaSnip', 'saadparwaiz1/cmp_luasnip' },
+--   }
+--
+--   use { -- Highlight, edit, and navigate code -- converted
+--     'nvim-treesitter/nvim-treesitter',
+--     run = function()
+--       pcall(require('nvim-treesitter.install').update { with_sync = true })
+--     end,
+--   }
+--
+--   use { -- Additional text objects via treesitter -- converted
+--     'nvim-treesitter/nvim-treesitter-textobjects',
+--     after = 'nvim-treesitter',
+--   }
+--
+--   -- Git related plugins - converted
+--   use 'tpope/vim-fugitive'
+--   use 'tpope/vim-rhubarb'
+--   use 'lewis6991/gitsigns.nvim'
+--
+--   -- Look good
+--   use 'nvim-lualine/lualine.nvim' -- Fancier statusline - converted
+--   use 'lukas-reineke/indent-blankline.nvim' -- Add indentation guides even on blank lines - converted
+--   use 'numToStr/Comment.nvim' -- "gc" to comment visual regions/lines - converted
+--   use 'tpope/vim-sleuth' -- Detect tabstop and shiftwidth automatically - dropping
+--   use { "catppuccin/nvim", as = "catppuccin" } -- converted
+--
+--   -- Go - converted
+--   use 'ray-x/go.nvim'
+--   use 'ray-x/guihua.lua'
+--
+--   -- File tree - converted
+--   use 'nvim-tree/nvim-web-devicons'
+--   use 'nvim-tree/nvim-tree.lua'
+--
+--   -- Fuzzy Finder (files, lsp, etc) - converted
+--   use { 'nvim-telescope/telescope.nvim', branch = '0.1.x', requires = { 'nvim-lua/plenary.nvim' } }
+--
+--   -- Fuzzy Finder Algorithm which requires local dependencies to be built. Only load if `make` is available - converted
+--   use { 'nvim-telescope/telescope-fzf-native.nvim', run = 'make', cond = vim.fn.executable 'make' == 1 }
+--
+--   use({ -- converted
+--       "aserowy/tmux.nvim",
+--       config = function() return require("tmux").setup() end
+--   })
+--
+--   -- Add custom plugins to packer from ~/.config/nvim/lua/custom/plugins.lua
+--   local has_plugins, plugins = pcall(require, 'custom.plugins')
+--   if has_plugins then
+--     plugins(use)
+--   end
+--
+--   if is_bootstrap then
+--     require('packer').sync()
+--   end
+-- end)
+--
+-- vim.g.loaded_netrw = 1
+-- vim.g.loaded_netrwPlugin = 1
+--
+--
+-- -- When we are bootstrapping a configuration, it doesn't
+-- -- make sense to execute the rest of the init.lua.
+-- --
+-- -- You'll need to restart nvim, and then it will work.
+-- if is_bootstrap then
+--   print '=================================='
+--   print '    Plugins are being installed'
+--   print '    Wait until Packer completes,'
+--   print '       then restart nvim'
+--   print '=================================='
+--   return
+-- end
+--
+-- -- Automatically source and re-compile packer whenever you save this init.lua
+-- local packer_group = vim.api.nvim_create_augroup('Packer', { clear = true })
+-- vim.api.nvim_create_autocmd('BufWritePost', {
+--   command = 'source <afile> | PackerCompile',
+--   group = packer_group,
+--   pattern = vim.fn.expand '$MYVIMRC',
+-- })
+--
+require("lazy").setup("plugins")
 -- [[ Setting options ]]
 -- See `:help vim.o`
 
@@ -143,6 +144,10 @@ vim.cmd.colorscheme "catppuccin"
 
 -- Set completeopt to have a better completion experience
 vim.o.completeopt = 'menuone,noselect'
+
+vim.o.tabstop = 2
+vim.o.shiftwidth = 2
+vim.o.shiftwidth = 2
 
 -- [[ Basic Keymaps ]]
 -- Set , as the leader key
@@ -186,523 +191,523 @@ vim.api.nvim_create_autocmd('TextYankPost', {
   pattern = '*',
 })
 
--- Catppuccin settings
-require("catppuccin").setup {
-  flavor = "mocha",
-  integrations = {
-    cmp = true,
-    gitsigns = true,
-    nvimtree = true,
-    telescope = true,
-  }
-}
-
--- Set lualine as statusline
--- See `:help lualine.txt`
-require('lualine').setup {
-  options = {
-    icons_enabled = false,
-    theme = 'catppuccin',
-    component_separators = '|',
-    section_separators = '',
-  },
-}
-
--- Enable Comment.nvim
-require('Comment').setup()
-
--- Enable `lukas-reineke/indent-blankline.nvim`
--- See `:help indent_blankline.txt`
-require('indent_blankline').setup {
-  char = '┊',
-  show_trailing_blankline_indent = false,
-}
-
--- Gitsigns
--- See `:help gitsigns.txt`
-require('gitsigns').setup {
-  signs = {
-    add = { text = '+' },
-    change = { text = '~' },
-    delete = { text = '_' },
-    topdelete = { text = '‾' },
-    changedelete = { text = '~' },
-  },
-}
-
--- [[ Configure Telescope ]]
--- See `:help telescope` and `:help telescope.setup()`
-require('telescope').setup {
-  defaults = {
-    mappings = {
-      i = {
-        ['<C-j>'] = require('telescope.actions').move_selection_next,
-        ['<C-k>'] = require('telescope.actions').move_selection_previous,
-      },
-      n = {
-        ['jk'] = 'close',
-      },
-    },
-  },
-}
-
--- Enable telescope fzf native, if installed
-pcall(require('telescope').load_extension, 'fzf')
-
--- See `:help telescope.builtin`
-vim.keymap.set('n', '<leader>?', require('telescope.builtin').oldfiles, { desc = '[?] Find recently opened files' })
-vim.keymap.set('n', '<leader><space>', require('telescope.builtin').buffers, { desc = '[ ] Find existing buffers' })
-vim.keymap.set('n', '<leader>/', function()
-  -- You can pass additional configuration to telescope to change theme, layout, etc.
-  require('telescope.builtin').current_buffer_fuzzy_find(require('telescope.themes').get_dropdown {
-    winblend = 0,
-    previewer = false,
-  })
-end, { desc = '[/] Fuzzily search in current buffer]' })
-
-vim.keymap.set('n', '<leader>sf', require('telescope.builtin').find_files, { desc = '[S]earch [F]iles' })
-vim.keymap.set('n', '<leader>sh', require('telescope.builtin').help_tags, { desc = '[S]earch [H]elp' })
-vim.keymap.set('n', '<leader>sw', require('telescope.builtin').grep_string, { desc = '[S]earch current [W]ord' })
-vim.keymap.set('n', '<leader>sg', require('telescope.builtin').live_grep, { desc = '[S]earch by [G]rep' })
-vim.keymap.set('n', '<leader>sd', require('telescope.builtin').diagnostics, { desc = '[S]earch [D]iagnostics' })
-vim.keymap.set('n', '<leader>km', require('telescope.builtin').keymaps, { desc = '[K]eymap [M]enu' })
-
--- empty setup using defaults
-local function on_attach(bufnr)
-  local api = require('nvim-tree.api')
-
-  local function opts(desc)
-    return { desc = 'nvim-tree: ' .. desc, buffer = bufnr, noremap = true, silent = true, nowait = true }
-  end
-
-
-  -- Default mappings. Feel free to modify or remove as you wish.
-  --
-  -- BEGIN_DEFAULT_ON_ATTACH
-  vim.keymap.set('n', '<C-]>', api.tree.change_root_to_node,          opts('CD'))
-  vim.keymap.set('n', '<C-e>', api.node.open.replace_tree_buffer,     opts('Open: In Place'))
-  vim.keymap.set('n', '<C-k>', api.node.show_info_popup,              opts('Info'))
-  vim.keymap.set('n', '<C-r>', api.fs.rename_sub,                     opts('Rename: Omit Filename'))
-  vim.keymap.set('n', '<C-t>', api.node.open.tab,                     opts('Open: New Tab'))
-  vim.keymap.set('n', '<C-v>', api.node.open.vertical,                opts('Open: Vertical Split'))
-  vim.keymap.set('n', '<C-x>', api.node.open.horizontal,              opts('Open: Horizontal Split'))
-  vim.keymap.set('n', '<BS>',  api.node.navigate.parent_close,        opts('Close Directory'))
-  vim.keymap.set('n', '<CR>',  api.node.open.edit,                    opts('Open'))
-  vim.keymap.set('n', '<Tab>', api.node.open.preview,                 opts('Open Preview'))
-  vim.keymap.set('n', '>',     api.node.navigate.sibling.next,        opts('Next Sibling'))
-  vim.keymap.set('n', '<',     api.node.navigate.sibling.prev,        opts('Previous Sibling'))
-  vim.keymap.set('n', '.',     api.node.run.cmd,                      opts('Run Command'))
-  vim.keymap.set('n', '-',     api.tree.change_root_to_parent,        opts('Up'))
-  vim.keymap.set('n', 'a',     api.fs.create,                         opts('Create'))
-  vim.keymap.set('n', 'bd',    api.marks.bulk.delete,                 opts('Delete Bookmarked'))
-  vim.keymap.set('n', 'bmv',   api.marks.bulk.move,                   opts('Move Bookmarked'))
-  vim.keymap.set('n', 'B',     api.tree.toggle_no_buffer_filter,      opts('Toggle No Buffer'))
-  vim.keymap.set('n', 'c',     api.fs.copy.node,                      opts('Copy'))
-  vim.keymap.set('n', 'C',     api.tree.toggle_git_clean_filter,      opts('Toggle Git Clean'))
-  vim.keymap.set('n', '[c',    api.node.navigate.git.prev,            opts('Prev Git'))
-  vim.keymap.set('n', ']c',    api.node.navigate.git.next,            opts('Next Git'))
-  vim.keymap.set('n', 'd',     api.fs.remove,                         opts('Delete'))
-  vim.keymap.set('n', 'D',     api.fs.trash,                          opts('Trash'))
-  vim.keymap.set('n', 'E',     api.tree.expand_all,                   opts('Expand All'))
-  vim.keymap.set('n', 'e',     api.fs.rename_basename,                opts('Rename: Basename'))
-  vim.keymap.set('n', ']e',    api.node.navigate.diagnostics.next,    opts('Next Diagnostic'))
-  vim.keymap.set('n', '[e',    api.node.navigate.diagnostics.prev,    opts('Prev Diagnostic'))
-  vim.keymap.set('n', 'F',     api.live_filter.clear,                 opts('Clean Filter'))
-  vim.keymap.set('n', 'f',     api.live_filter.start,                 opts('Filter'))
-  vim.keymap.set('n', 'g?',    api.tree.toggle_help,                  opts('Help'))
-  vim.keymap.set('n', 'gy',    api.fs.copy.absolute_path,             opts('Copy Absolute Path'))
-  vim.keymap.set('n', 'H',     api.tree.toggle_hidden_filter,         opts('Toggle Dotfiles'))
-  vim.keymap.set('n', 'I',     api.tree.toggle_gitignore_filter,      opts('Toggle Git Ignore'))
-  vim.keymap.set('n', 'J',     api.node.navigate.sibling.last,        opts('Last Sibling'))
-  vim.keymap.set('n', 'K',     api.node.navigate.sibling.first,       opts('First Sibling'))
-  vim.keymap.set('n', 'm',     api.marks.toggle,                      opts('Toggle Bookmark'))
-  vim.keymap.set('n', 'o',     api.node.open.edit,                    opts('Open'))
-  vim.keymap.set('n', 'O',     api.node.open.no_window_picker,        opts('Open: No Window Picker'))
-  vim.keymap.set('n', 'p',     api.fs.paste,                          opts('Paste'))
-  vim.keymap.set('n', 'P',     api.node.navigate.parent,              opts('Parent Directory'))
-  vim.keymap.set('n', 'q',     api.tree.close,                        opts('Close'))
-  vim.keymap.set('n', 'r',     api.fs.rename,                         opts('Rename'))
-  vim.keymap.set('n', 'R',     api.tree.reload,                       opts('Refresh'))
-  vim.keymap.set('n', 's',     api.node.run.system,                   opts('Run System'))
-  vim.keymap.set('n', 'S',     api.tree.search_node,                  opts('Search'))
-  vim.keymap.set('n', 'U',     api.tree.toggle_custom_filter,         opts('Toggle Hidden'))
-  vim.keymap.set('n', 'W',     api.tree.collapse_all,                 opts('Collapse'))
-  vim.keymap.set('n', 'x',     api.fs.cut,                            opts('Cut'))
-  vim.keymap.set('n', 'y',     api.fs.copy.filename,                  opts('Copy Name'))
-  vim.keymap.set('n', 'Y',     api.fs.copy.relative_path,             opts('Copy Relative Path'))
-  vim.keymap.set('n', '<2-LeftMouse>',  api.node.open.edit,           opts('Open'))
-  vim.keymap.set('n', '<2-RightMouse>', api.tree.change_root_to_node, opts('CD'))
-  -- END_DEFAULT_ON_ATTACH
-
-
-  -- Mappings migrated from view.mappings.list
-  --
-  -- You will need to insert "your code goes here" for any mappings with a custom action_cb
-  vim.keymap.set('n', 'u', api.tree.change_root_to_parent, opts('Up'))
-end
-
-require("nvim-tree").setup({
-  sort_by = "case_sensitive",
-  on_attach = on_attach,
-  view = {
-    adaptive_size = true,
-  },
-  hijack_cursor = true,
-  hijack_unnamed_buffer_when_opening = false,
-  sync_root_with_cwd = true,
-  update_focused_file = {
-    enable = true,
-    update_root = false,
-  },
-  git = {
-    enable = false,
-    ignore = true,
-  },
-  filesystem_watchers = {
-    enable = true,
-  },
-  renderer = {
-    root_folder_label = false,
-    highlight_git = false,
-    highlight_opened_files = "none",
-    indent_markers = {
-      enable = false,
-    },
-    icons = {
-      show = {
-        file = true,
-        folder = true,
-        folder_arrow = true,
-        git = false,
-      },
-      glyphs = {
-        default = "󰈚",
-        symlink = "",
-        folder = {
-          default = "",
-          empty = "",
-          empty_open = "",
-          open = "",
-          symlink = "",
-          symlink_open = "",
-          arrow_open = "",
-          arrow_closed = "",
-        },
-        git = {
-          unstaged = "✗",
-          staged = "✓",
-          unmerged = "",
-          renamed = "➜",
-          untracked = "★",
-          deleted = "",
-          ignored = "◌",
-        },
-      },
-    },
-  },
-  filters = {
-    dotfiles = false,
-  },
-})
-
--- mappings
-vim.keymap.set('n', '<C-n>', ':NvimTreeToggle<CR>')
-vim.keymap.set('n', '<leader>m', ':NvimTreeFindFile<CR>')
-
--- [[ Configure Treesitter ]]
--- See `:help nvim-treesitter`
-require('nvim-treesitter.configs').setup {
-  -- Add languages to be installed here that you want installed for treesitter
-  ensure_installed = { 'c', 'cpp', 'go', 'lua', 'vim', 'python', 'rust', 'typescript', 'hcl', 'javascript', 'ruby', 'dockerfile', 'c_sharp' },
-
-  highlight = { enable = true },
-  indent = { enable = true, disable = { 'python' } },
-  incremental_selection = {
-    enable = true,
-    keymaps = {
-      init_selection = '<c-space>',
-      node_incremental = '<c-space>',
-      scope_incremental = '<c-s>',
-      node_decremental = '<c-backspace>',
-    },
-  },
-  textobjects = {
-    select = {
-      enable = true,
-      lookahead = true, -- Automatically jump forward to textobj, similar to targets.vim
-      keymaps = {
-        -- You can use the capture groups defined in textobjects.scm
-        ['aa'] = '@parameter.outer',
-        ['ia'] = '@parameter.inner',
-        ['af'] = '@function.outer',
-        ['if'] = '@function.inner',
-        ['ac'] = '@class.outer',
-        ['ic'] = '@class.inner',
-      },
-    },
-    move = {
-      enable = true,
-      set_jumps = true, -- whether to set jumps in the jumplist
-      goto_next_start = {
-        [']m'] = '@function.outer',
-        [']]'] = '@class.outer',
-      },
-      goto_next_end = {
-        [']M'] = '@function.outer',
-        [']['] = '@class.outer',
-      },
-      goto_previous_start = {
-        ['[m'] = '@function.outer',
-        ['[['] = '@class.outer',
-      },
-      goto_previous_end = {
-        ['[M'] = '@function.outer',
-        ['[]'] = '@class.outer',
-      },
-    },
-    swap = {
-      enable = true,
-      swap_next = {
-        ['<leader>a'] = '@parameter.inner',
-      },
-      swap_previous = {
-        ['<leader>A'] = '@parameter.inner',
-      },
-    },
-  },
-}
-
--- Diagnostic keymaps
-vim.keymap.set('n', '[d', vim.diagnostic.goto_prev)
-vim.keymap.set('n', ']d', vim.diagnostic.goto_next)
-vim.keymap.set('n', '<leader>e', vim.diagnostic.open_float)
-vim.keymap.set('n', '<leader>q', vim.diagnostic.setloclist)
-
--- LSP settings.
---  This function gets run when an LSP connects to a particular buffer.
-local on_attach = function(_, bufnr)
-  -- NOTE: Remember that lua is a real programming language, and as such it is possible
-  -- to define small helper and utility functions so you don't have to repeat yourself
-  -- many times.
-  --
-  -- In this case, we create a function that lets us more easily define mappings specific
-  -- for LSP related items. It sets the mode, buffer and description for us each time.
-  local nmap = function(keys, func, desc)
-    if desc then
-      desc = 'LSP: ' .. desc
-    end
-
-    vim.keymap.set('n', keys, func, { buffer = bufnr, desc = desc })
-  end
-
-  nmap('<leader>rn', vim.lsp.buf.rename, '[R]e[n]ame')
-  nmap('<leader>ca', vim.lsp.buf.code_action, '[C]ode [A]ction')
-
-  nmap('gd', vim.lsp.buf.definition, '[G]oto [D]efinition')
-  nmap('gr', require('telescope.builtin').lsp_references, '[G]oto [R]eferences')
-  nmap('gI', vim.lsp.buf.implementation, '[G]oto [I]mplementation')
-  nmap('<leader>D', vim.lsp.buf.type_definition, 'Type [D]efinition')
-  nmap('<leader>ds', require('telescope.builtin').lsp_document_symbols, '[D]ocument [S]ymbols')
-  nmap('<leader>ws', require('telescope.builtin').lsp_dynamic_workspace_symbols, '[W]orkspace [S]ymbols')
-
-  -- See `:help K` for why this keymap
-  nmap('K', vim.lsp.buf.hover, 'Hover Documentation')
-  nmap('<C-p>', vim.lsp.buf.signature_help, 'Signature Documentation')
-
-  -- Lesser used LSP functionality
-  nmap('gD', vim.lsp.buf.declaration, '[G]oto [D]eclaration')
-  nmap('<leader>wa', vim.lsp.buf.add_workspace_folder, '[W]orkspace [A]dd Folder')
-  nmap('<leader>wr', vim.lsp.buf.remove_workspace_folder, '[W]orkspace [R]emove Folder')
-  nmap('<leader>wl', function()
-    print(vim.inspect(vim.lsp.buf.list_workspace_folders()))
-  end, '[W]orkspace [L]ist Folders')
-
-  -- Create a command `:Format` local to the LSP buffer
-  vim.api.nvim_buf_create_user_command(bufnr, 'Format', function(_)
-    vim.lsp.buf.format()
-  end, { desc = 'Format current buffer with LSP' })
-end
-
--- Go goodies
-require('go').setup({
-    -- notify: use nvim-notify
-    notify = false,
-    -- auto commands
-    auto_format = true,
-    auto_lint = true,
-    -- linters: revive, errcheck, staticcheck, golangci-lint
-    linter = 'golangci-lint',
-    -- linter_flags: e.g., {revive = {'-config', '/path/to/config.yml'}}
-    linter_flags = {},
-    -- lint_prompt_style: qf (quickfix), vt (virtual text)
-    lint_prompt_style = 'qf',
-    -- formatter: goimports, gofmt, gofumpt, lsp
-    formatter = 'goimports',
-    -- maintain cursor position after formatting loaded buffer
-    maintain_cursor_pos = false,
-    -- test flags: -count=1 will disable cache
-    test_flags = {'-v'},
-    test_timeout = '30s',
-    test_env = {},
-    -- show test result with popup window
-    test_popup = true,
-    test_popup_auto_leave = false,
-    test_popup_width = 80,
-    test_popup_height = 10,
-    -- test open
-    test_open_cmd = 'edit',
-    -- struct tags
-    tags_name = 'json',
-    tags_options = {'json=omitempty'},
-    tags_transform = 'snakecase',
-    tags_flags = {'-skip-unexported'},
-    -- quick type
-    quick_type_flags = {'--just-types'},
-})
-
--- Enable the following language servers
---  Feel free to add/remove any LSPs that you want here. They will automatically be installed.
+-- -- Catppuccin settings
+-- require("catppuccin").setup {
+--   flavor = "mocha",
+--   integrations = {
+--     cmp = true,
+--     gitsigns = true,
+--     nvimtree = true,
+--     telescope = true,
+--   }
+-- }
 --
---  Add any additional override configuration in the following tables. They will be passed to
---  the `settings` field of the server config. You must look up that documentation yourself.
-local servers = {
-  cssls = {
-    css = {
-      validate = true,
-    },
-    less = {
-      validate = true,
-    },
-    scss = {
-      validate = true,
-    },
-  },
-  gopls = {
-    analyses = {
-			unusedparams = true,
-		},
-		staticcheck = true,
-		linksInHover = false,
-		codelenses = {
-			generate = true,
-			gc_details = true,
-			regenerate_cgo = true,
-			tidy = true,
-			upgrade_depdendency = true,
-			vendor = true,
-		},
-		usePlaceholders = true,
-  },
-  lua_ls = {
-    Lua = {
-      runtime = {
-        -- Tell the language server which version of Lua you're using (most likely LuaJIT in the case of Neovim)
-        version = 'LuaJIT',
-      },
-      diagnostics = {
-        -- Get the language server to recognize the `vim` global
-        globals = {'vim'},
-      },
-      workspace = {
-        -- Make the server aware of Neovim runtime files
-        library = vim.api.nvim_get_runtime_file("", true),
-        checkThirdParty = false,
-      },
-      -- Do not send telemetry data containing a randomized but unique identifier
-      telemetry = {
-        enable = false,
-      },
-    },
-  },
-  rust_analyzer = {},
-  tsserver = {
-    filetypes = {
-      'javascript',
-      'javascriptreact',
-      'javascript.jsx',
-      'typescript',
-      'typescriptreact',
-      'typescript.tsx',
-    },
-    init_options = {
-      hostInfo = 'neovim',
-    },
-  },
-}
-
--- Setup neovim lua configuration
-require('neodev').setup()
-
--- Setup 
--- nvim-cmp supports additional completion capabilities, so broadcast that to servers
-local capabilities = vim.lsp.protocol.make_client_capabilities()
-capabilities = require('cmp_nvim_lsp').default_capabilities(capabilities)
-
--- Setup mason so it can manage external tooling
-require('mason').setup()
-
--- Ensure the servers above are installed
-local mason_lspconfig = require 'mason-lspconfig'
-
-mason_lspconfig.setup {
-  ensure_installed = vim.tbl_keys(servers),
-}
-
-mason_lspconfig.setup_handlers {
-  function(server_name)
-    require('lspconfig')[server_name].setup {
-      capabilities = capabilities,
-      on_attach = on_attach,
-      settings = servers[server_name],
-    }
-  end,
-}
-
--- Turn on lsp status information
-require('fidget').setup()
-
--- nvim-cmp setup
-local cmp = require 'cmp'
-local luasnip = require 'luasnip'
-
-cmp.setup {
-  snippet = {
-    expand = function(args)
-      luasnip.lsp_expand(args.body)
-    end,
-  },
-  mapping = cmp.mapping.preset.insert {
-    ['<C-d>'] = cmp.mapping.scroll_docs(-4),
-    ['<C-f>'] = cmp.mapping.scroll_docs(4),
-    ['<C-Space>'] = cmp.mapping.complete(),
-    ['<CR>'] = cmp.mapping.confirm {
-      behavior = cmp.ConfirmBehavior.Replace,
-      select = true,
-    },
-    ['<C-j>'] = cmp.mapping(function(fallback)
-      if cmp.visible() then
-        cmp.select_next_item()
-      elseif luasnip.expand_or_jumpable() then
-        luasnip.expand_or_jump()
-      else
-        fallback()
-      end
-    end, { 'i', 's' }),
-    ['<C-k>'] = cmp.mapping(function(fallback)
-      if cmp.visible() then
-        cmp.select_prev_item()
-      elseif luasnip.jumpable(-1) then
-        luasnip.jump(-1)
-      else
-        fallback()
-      end
-    end, { 'i', 's' }),
-  },
-  sources = {
-    { name = 'nvim_lsp' },
-    { name = 'luasnip' },
-  },
-}
-
+-- -- Set lualine as statusline
+-- -- See `:help lualine.txt`
+-- require('lualine').setup {
+--   options = {
+--     icons_enabled = false,
+--     theme = 'catppuccin',
+--     component_separators = '|',
+--     section_separators = '',
+--   },
+-- }
+--
+-- -- Enable Comment.nvim
+-- require('Comment').setup()
+--
+-- -- Enable `lukas-reineke/indent-blankline.nvim`
+-- -- See `:help indent_blankline.txt`
+-- require('indent_blankline').setup {
+--   char = '┊',
+--   show_trailing_blankline_indent = false,
+-- }
+--
+-- -- Gitsigns
+-- -- See `:help gitsigns.txt`
+-- require('gitsigns').setup {
+--   signs = {
+--     add = { text = '+' },
+--     change = { text = '~' },
+--     delete = { text = '_' },
+--     topdelete = { text = '‾' },
+--     changedelete = { text = '~' },
+--   },
+-- }
+--
+-- -- [[ Configure Telescope ]]
+-- -- See `:help telescope` and `:help telescope.setup()`
+-- require('telescope').setup {
+--   defaults = {
+--     mappings = {
+--       i = {
+--         ['<C-j>'] = require('telescope.actions').move_selection_next,
+--         ['<C-k>'] = require('telescope.actions').move_selection_previous,
+--       },
+--       n = {
+--         ['jk'] = 'close',
+--       },
+--     },
+--   },
+-- }
+--
+-- -- Enable telescope fzf native, if installed
+-- pcall(require('telescope').load_extension, 'fzf')
+--
+-- -- See `:help telescope.builtin`
+-- vim.keymap.set('n', '<leader>?', require('telescope.builtin').oldfiles, { desc = '[?] Find recently opened files' })
+-- vim.keymap.set('n', '<leader><space>', require('telescope.builtin').buffers, { desc = '[ ] Find existing buffers' })
+-- vim.keymap.set('n', '<leader>/', function()
+--   -- You can pass additional configuration to telescope to change theme, layout, etc.
+--   require('telescope.builtin').current_buffer_fuzzy_find(require('telescope.themes').get_dropdown {
+--     winblend = 0,
+--     previewer = false,
+--   })
+-- end, { desc = '[/] Fuzzily search in current buffer]' })
+--
+-- vim.keymap.set('n', '<leader>sf', require('telescope.builtin').find_files, { desc = '[S]earch [F]iles' })
+-- vim.keymap.set('n', '<leader>sh', require('telescope.builtin').help_tags, { desc = '[S]earch [H]elp' })
+-- vim.keymap.set('n', '<leader>sw', require('telescope.builtin').grep_string, { desc = '[S]earch current [W]ord' })
+-- vim.keymap.set('n', '<leader>sg', require('telescope.builtin').live_grep, { desc = '[S]earch by [G]rep' })
+-- vim.keymap.set('n', '<leader>sd', require('telescope.builtin').diagnostics, { desc = '[S]earch [D]iagnostics' })
+-- vim.keymap.set('n', '<leader>km', require('telescope.builtin').keymaps, { desc = '[K]eymap [M]enu' })
+--
+-- -- empty setup using defaults
+-- local function on_attach(bufnr)
+--   local api = require('nvim-tree.api')
+--
+--   local function opts(desc)
+--     return { desc = 'nvim-tree: ' .. desc, buffer = bufnr, noremap = true, silent = true, nowait = true }
+--   end
+--
+--
+--   -- Default mappings. Feel free to modify or remove as you wish.
+--   --
+--   -- BEGIN_DEFAULT_ON_ATTACH
+--   vim.keymap.set('n', '<C-]>', api.tree.change_root_to_node,          opts('CD'))
+--   vim.keymap.set('n', '<C-e>', api.node.open.replace_tree_buffer,     opts('Open: In Place'))
+--   vim.keymap.set('n', '<C-k>', api.node.show_info_popup,              opts('Info'))
+--   vim.keymap.set('n', '<C-r>', api.fs.rename_sub,                     opts('Rename: Omit Filename'))
+--   vim.keymap.set('n', '<C-t>', api.node.open.tab,                     opts('Open: New Tab'))
+--   vim.keymap.set('n', '<C-v>', api.node.open.vertical,                opts('Open: Vertical Split'))
+--   vim.keymap.set('n', '<C-x>', api.node.open.horizontal,              opts('Open: Horizontal Split'))
+--   vim.keymap.set('n', '<BS>',  api.node.navigate.parent_close,        opts('Close Directory'))
+--   vim.keymap.set('n', '<CR>',  api.node.open.edit,                    opts('Open'))
+--   vim.keymap.set('n', '<Tab>', api.node.open.preview,                 opts('Open Preview'))
+--   vim.keymap.set('n', '>',     api.node.navigate.sibling.next,        opts('Next Sibling'))
+--   vim.keymap.set('n', '<',     api.node.navigate.sibling.prev,        opts('Previous Sibling'))
+--   vim.keymap.set('n', '.',     api.node.run.cmd,                      opts('Run Command'))
+--   vim.keymap.set('n', '-',     api.tree.change_root_to_parent,        opts('Up'))
+--   vim.keymap.set('n', 'a',     api.fs.create,                         opts('Create'))
+--   vim.keymap.set('n', 'bd',    api.marks.bulk.delete,                 opts('Delete Bookmarked'))
+--   vim.keymap.set('n', 'bmv',   api.marks.bulk.move,                   opts('Move Bookmarked'))
+--   vim.keymap.set('n', 'B',     api.tree.toggle_no_buffer_filter,      opts('Toggle No Buffer'))
+--   vim.keymap.set('n', 'c',     api.fs.copy.node,                      opts('Copy'))
+--   vim.keymap.set('n', 'C',     api.tree.toggle_git_clean_filter,      opts('Toggle Git Clean'))
+--   vim.keymap.set('n', '[c',    api.node.navigate.git.prev,            opts('Prev Git'))
+--   vim.keymap.set('n', ']c',    api.node.navigate.git.next,            opts('Next Git'))
+--   vim.keymap.set('n', 'd',     api.fs.remove,                         opts('Delete'))
+--   vim.keymap.set('n', 'D',     api.fs.trash,                          opts('Trash'))
+--   vim.keymap.set('n', 'E',     api.tree.expand_all,                   opts('Expand All'))
+--   vim.keymap.set('n', 'e',     api.fs.rename_basename,                opts('Rename: Basename'))
+--   vim.keymap.set('n', ']e',    api.node.navigate.diagnostics.next,    opts('Next Diagnostic'))
+--   vim.keymap.set('n', '[e',    api.node.navigate.diagnostics.prev,    opts('Prev Diagnostic'))
+--   vim.keymap.set('n', 'F',     api.live_filter.clear,                 opts('Clean Filter'))
+--   vim.keymap.set('n', 'f',     api.live_filter.start,                 opts('Filter'))
+--   vim.keymap.set('n', 'g?',    api.tree.toggle_help,                  opts('Help'))
+--   vim.keymap.set('n', 'gy',    api.fs.copy.absolute_path,             opts('Copy Absolute Path'))
+--   vim.keymap.set('n', 'H',     api.tree.toggle_hidden_filter,         opts('Toggle Dotfiles'))
+--   vim.keymap.set('n', 'I',     api.tree.toggle_gitignore_filter,      opts('Toggle Git Ignore'))
+--   vim.keymap.set('n', 'J',     api.node.navigate.sibling.last,        opts('Last Sibling'))
+--   vim.keymap.set('n', 'K',     api.node.navigate.sibling.first,       opts('First Sibling'))
+--   vim.keymap.set('n', 'm',     api.marks.toggle,                      opts('Toggle Bookmark'))
+--   vim.keymap.set('n', 'o',     api.node.open.edit,                    opts('Open'))
+--   vim.keymap.set('n', 'O',     api.node.open.no_window_picker,        opts('Open: No Window Picker'))
+--   vim.keymap.set('n', 'p',     api.fs.paste,                          opts('Paste'))
+--   vim.keymap.set('n', 'P',     api.node.navigate.parent,              opts('Parent Directory'))
+--   vim.keymap.set('n', 'q',     api.tree.close,                        opts('Close'))
+--   vim.keymap.set('n', 'r',     api.fs.rename,                         opts('Rename'))
+--   vim.keymap.set('n', 'R',     api.tree.reload,                       opts('Refresh'))
+--   vim.keymap.set('n', 's',     api.node.run.system,                   opts('Run System'))
+--   vim.keymap.set('n', 'S',     api.tree.search_node,                  opts('Search'))
+--   vim.keymap.set('n', 'U',     api.tree.toggle_custom_filter,         opts('Toggle Hidden'))
+--   vim.keymap.set('n', 'W',     api.tree.collapse_all,                 opts('Collapse'))
+--   vim.keymap.set('n', 'x',     api.fs.cut,                            opts('Cut'))
+--   vim.keymap.set('n', 'y',     api.fs.copy.filename,                  opts('Copy Name'))
+--   vim.keymap.set('n', 'Y',     api.fs.copy.relative_path,             opts('Copy Relative Path'))
+--   vim.keymap.set('n', '<2-LeftMouse>',  api.node.open.edit,           opts('Open'))
+--   vim.keymap.set('n', '<2-RightMouse>', api.tree.change_root_to_node, opts('CD'))
+--   -- END_DEFAULT_ON_ATTACH
+--
+--
+--   -- Mappings migrated from view.mappings.list
+--   --
+--   -- You will need to insert "your code goes here" for any mappings with a custom action_cb
+--   vim.keymap.set('n', 'u', api.tree.change_root_to_parent, opts('Up'))
+-- end
+--
+-- require("nvim-tree").setup({
+--   sort_by = "case_sensitive",
+--   on_attach = on_attach,
+--   view = {
+--     adaptive_size = true,
+--   },
+--   hijack_cursor = true,
+--   hijack_unnamed_buffer_when_opening = false,
+--   sync_root_with_cwd = true,
+--   update_focused_file = {
+--     enable = true,
+--     update_root = false,
+--   },
+--   git = {
+--     enable = false,
+--     ignore = true,
+--   },
+--   filesystem_watchers = {
+--     enable = true,
+--   },
+--   renderer = {
+--     root_folder_label = false,
+--     highlight_git = false,
+--     highlight_opened_files = "none",
+--     indent_markers = {
+--       enable = false,
+--     },
+--     icons = {
+--       show = {
+--         file = true,
+--         folder = true,
+--         folder_arrow = true,
+--         git = false,
+--       },
+--       glyphs = {
+--         default = "󰈚",
+--         symlink = "",
+--         folder = {
+--           default = "",
+--           empty = "",
+--           empty_open = "",
+--           open = "",
+--           symlink = "",
+--           symlink_open = "",
+--           arrow_open = "",
+--           arrow_closed = "",
+--         },
+--         git = {
+--           unstaged = "✗",
+--           staged = "✓",
+--           unmerged = "",
+--           renamed = "➜",
+--           untracked = "★",
+--           deleted = "",
+--           ignored = "◌",
+--         },
+--       },
+--     },
+--   },
+--   filters = {
+--     dotfiles = false,
+--   },
+-- })
+--
+-- -- mappings
+-- vim.keymap.set('n', '<C-n>', ':NvimTreeToggle<CR>')
+-- vim.keymap.set('n', '<leader>m', ':NvimTreeFindFile<CR>')
+--
+-- -- [[ Configure Treesitter ]]
+-- -- See `:help nvim-treesitter`
+-- require('nvim-treesitter.configs').setup {
+--   -- Add languages to be installed here that you want installed for treesitter
+--   ensure_installed = { 'c', 'cpp', 'go', 'lua', 'vim', 'python', 'rust', 'typescript', 'hcl', 'javascript', 'ruby', 'dockerfile', 'c_sharp' },
+--
+--   highlight = { enable = true },
+--   indent = { enable = true, disable = { 'python' } },
+--   incremental_selection = {
+--     enable = true,
+--     keymaps = {
+--       init_selection = '<c-space>',
+--       node_incremental = '<c-space>',
+--       scope_incremental = '<c-s>',
+--       node_decremental = '<c-backspace>',
+--     },
+--   },
+--   textobjects = {
+--     select = {
+--       enable = true,
+--       lookahead = true, -- Automatically jump forward to textobj, similar to targets.vim
+--       keymaps = {
+--         -- You can use the capture groups defined in textobjects.scm
+--         ['aa'] = '@parameter.outer',
+--         ['ia'] = '@parameter.inner',
+--         ['af'] = '@function.outer',
+--         ['if'] = '@function.inner',
+--         ['ac'] = '@class.outer',
+--         ['ic'] = '@class.inner',
+--       },
+--     },
+--     move = {
+--       enable = true,
+--       set_jumps = true, -- whether to set jumps in the jumplist
+--       goto_next_start = {
+--         [']m'] = '@function.outer',
+--         [']]'] = '@class.outer',
+--       },
+--       goto_next_end = {
+--         [']M'] = '@function.outer',
+--         [']['] = '@class.outer',
+--       },
+--       goto_previous_start = {
+--         ['[m'] = '@function.outer',
+--         ['[['] = '@class.outer',
+--       },
+--       goto_previous_end = {
+--         ['[M'] = '@function.outer',
+--         ['[]'] = '@class.outer',
+--       },
+--     },
+--     swap = {
+--       enable = true,
+--       swap_next = {
+--         ['<leader>a'] = '@parameter.inner',
+--       },
+--       swap_previous = {
+--         ['<leader>A'] = '@parameter.inner',
+--       },
+--     },
+--   },
+-- }
+--
+-- -- Diagnostic keymaps
+-- vim.keymap.set('n', '[d', vim.diagnostic.goto_prev)
+-- vim.keymap.set('n', ']d', vim.diagnostic.goto_next)
+-- vim.keymap.set('n', '<leader>e', vim.diagnostic.open_float)
+-- vim.keymap.set('n', '<leader>q', vim.diagnostic.setloclist)
+--
+-- -- LSP settings.
+-- --  This function gets run when an LSP connects to a particular buffer.
+-- local on_attach = function(_, bufnr)
+--   -- NOTE: Remember that lua is a real programming language, and as such it is possible
+--   -- to define small helper and utility functions so you don't have to repeat yourself
+--   -- many times.
+--   --
+--   -- In this case, we create a function that lets us more easily define mappings specific
+--   -- for LSP related items. It sets the mode, buffer and description for us each time.
+--   local nmap = function(keys, func, desc)
+--     if desc then
+--       desc = 'LSP: ' .. desc
+--     end
+--
+--     vim.keymap.set('n', keys, func, { buffer = bufnr, desc = desc })
+--   end
+--
+--   nmap('<leader>rn', vim.lsp.buf.rename, '[R]e[n]ame')
+--   nmap('<leader>ca', vim.lsp.buf.code_action, '[C]ode [A]ction')
+--
+--   nmap('gd', vim.lsp.buf.definition, '[G]oto [D]efinition')
+--   nmap('gr', require('telescope.builtin').lsp_references, '[G]oto [R]eferences')
+--   nmap('gI', vim.lsp.buf.implementation, '[G]oto [I]mplementation')
+--   nmap('<leader>D', vim.lsp.buf.type_definition, 'Type [D]efinition')
+--   nmap('<leader>ds', require('telescope.builtin').lsp_document_symbols, '[D]ocument [S]ymbols')
+--   nmap('<leader>ws', require('telescope.builtin').lsp_dynamic_workspace_symbols, '[W]orkspace [S]ymbols')
+--
+--   -- See `:help K` for why this keymap
+--   nmap('K', vim.lsp.buf.hover, 'Hover Documentation')
+--   nmap('<C-p>', vim.lsp.buf.signature_help, 'Signature Documentation')
+--
+--   -- Lesser used LSP functionality
+--   nmap('gD', vim.lsp.buf.declaration, '[G]oto [D]eclaration')
+--   nmap('<leader>wa', vim.lsp.buf.add_workspace_folder, '[W]orkspace [A]dd Folder')
+--   nmap('<leader>wr', vim.lsp.buf.remove_workspace_folder, '[W]orkspace [R]emove Folder')
+--   nmap('<leader>wl', function()
+--     print(vim.inspect(vim.lsp.buf.list_workspace_folders()))
+--   end, '[W]orkspace [L]ist Folders')
+--
+--   -- Create a command `:Format` local to the LSP buffer
+--   vim.api.nvim_buf_create_user_command(bufnr, 'Format', function(_)
+--     vim.lsp.buf.format()
+--   end, { desc = 'Format current buffer with LSP' })
+-- end
+--
+-- -- Go goodies
+-- require('go').setup({
+--     -- notify: use nvim-notify
+--     notify = false,
+--     -- auto commands
+--     auto_format = true,
+--     auto_lint = true,
+--     -- linters: revive, errcheck, staticcheck, golangci-lint
+--     linter = 'golangci-lint',
+--     -- linter_flags: e.g., {revive = {'-config', '/path/to/config.yml'}}
+--     linter_flags = {},
+--     -- lint_prompt_style: qf (quickfix), vt (virtual text)
+--     lint_prompt_style = 'qf',
+--     -- formatter: goimports, gofmt, gofumpt, lsp
+--     formatter = 'goimports',
+--     -- maintain cursor position after formatting loaded buffer
+--     maintain_cursor_pos = false,
+--     -- test flags: -count=1 will disable cache
+--     test_flags = {'-v'},
+--     test_timeout = '30s',
+--     test_env = {},
+--     -- show test result with popup window
+--     test_popup = true,
+--     test_popup_auto_leave = false,
+--     test_popup_width = 80,
+--     test_popup_height = 10,
+--     -- test open
+--     test_open_cmd = 'edit',
+--     -- struct tags
+--     tags_name = 'json',
+--     tags_options = {'json=omitempty'},
+--     tags_transform = 'snakecase',
+--     tags_flags = {'-skip-unexported'},
+--     -- quick type
+--     quick_type_flags = {'--just-types'},
+-- })
+--
+-- -- Enable the following language servers
+-- --  Feel free to add/remove any LSPs that you want here. They will automatically be installed.
+-- --
+-- --  Add any additional override configuration in the following tables. They will be passed to
+-- --  the `settings` field of the server config. You must look up that documentation yourself.
+-- local servers = {
+--   cssls = {
+--     css = {
+--       validate = true,
+--     },
+--     less = {
+--       validate = true,
+--     },
+--     scss = {
+--       validate = true,
+--     },
+--   },
+--   gopls = {
+--     analyses = {
+-- 			unusedparams = true,
+-- 		},
+-- 		staticcheck = true,
+-- 		linksInHover = false,
+-- 		codelenses = {
+-- 			generate = true,
+-- 			gc_details = true,
+-- 			regenerate_cgo = true,
+-- 			tidy = true,
+-- 			upgrade_depdendency = true,
+-- 			vendor = true,
+-- 		},
+-- 		usePlaceholders = true,
+--   },
+--   lua_ls = {
+--     Lua = {
+--       runtime = {
+--         -- Tell the language server which version of Lua you're using (most likely LuaJIT in the case of Neovim)
+--         version = 'LuaJIT',
+--       },
+--       diagnostics = {
+--         -- Get the language server to recognize the `vim` global
+--         globals = {'vim'},
+--       },
+--       workspace = {
+--         -- Make the server aware of Neovim runtime files
+--         library = vim.api.nvim_get_runtime_file("", true),
+--         checkThirdParty = false,
+--       },
+--       -- Do not send telemetry data containing a randomized but unique identifier
+--       telemetry = {
+--         enable = false,
+--       },
+--     },
+--   },
+--   rust_analyzer = {},
+--   tsserver = {
+--     filetypes = {
+--       'javascript',
+--       'javascriptreact',
+--       'javascript.jsx',
+--       'typescript',
+--       'typescriptreact',
+--       'typescript.tsx',
+--     },
+--     init_options = {
+--       hostInfo = 'neovim',
+--     },
+--   },
+-- }
+--
+-- -- Setup neovim lua configuration
+-- require('neodev').setup()
+--
+-- -- Setup 
+-- -- nvim-cmp supports additional completion capabilities, so broadcast that to servers
+-- local capabilities = vim.lsp.protocol.make_client_capabilities()
+-- capabilities = require('cmp_nvim_lsp').default_capabilities(capabilities)
+--
+-- -- Setup mason so it can manage external tooling
+-- require('mason').setup()
+--
+-- -- Ensure the servers above are installed
+-- local mason_lspconfig = require 'mason-lspconfig'
+--
+-- mason_lspconfig.setup {
+--   ensure_installed = vim.tbl_keys(servers),
+-- }
+--
+-- mason_lspconfig.setup_handlers {
+--   function(server_name)
+--     require('lspconfig')[server_name].setup {
+--       capabilities = capabilities,
+--       on_attach = on_attach,
+--       settings = servers[server_name],
+--     }
+--   end,
+-- }
+--
+-- -- Turn on lsp status information
+-- require('fidget').setup()
+--
+-- -- nvim-cmp setup
+-- local cmp = require 'cmp'
+-- local luasnip = require 'luasnip'
+--
+-- cmp.setup {
+--   snippet = {
+--     expand = function(args)
+--       luasnip.lsp_expand(args.body)
+--     end,
+--   },
+--   mapping = cmp.mapping.preset.insert {
+--     ['<C-d>'] = cmp.mapping.scroll_docs(-4),
+--     ['<C-f>'] = cmp.mapping.scroll_docs(4),
+--     ['<C-Space>'] = cmp.mapping.complete(),
+--     ['<CR>'] = cmp.mapping.confirm {
+--       behavior = cmp.ConfirmBehavior.Replace,
+--       select = true,
+--     },
+--     ['<C-j>'] = cmp.mapping(function(fallback)
+--       if cmp.visible() then
+--         cmp.select_next_item()
+--       elseif luasnip.expand_or_jumpable() then
+--         luasnip.expand_or_jump()
+--       else
+--         fallback()
+--       end
+--     end, { 'i', 's' }),
+--     ['<C-k>'] = cmp.mapping(function(fallback)
+--       if cmp.visible() then
+--         cmp.select_prev_item()
+--       elseif luasnip.jumpable(-1) then
+--         luasnip.jump(-1)
+--       else
+--         fallback()
+--       end
+--     end, { 'i', 's' }),
+--   },
+--   sources = {
+--     { name = 'nvim_lsp' },
+--     { name = 'luasnip' },
+--   },
+-- }
+--
 -- The line beneath this is called `modeline`. See `:help modeline`
 -- vim: ts=2 sts=2 sw=2 et
