+++ dot_config/nvim/lua/plugins/catppuccin.lua
@@ -1,0 +1,18 @@
+return {
+	{
+		"catppuccin/nvim",
+		name = "catppuccin",
+		priority = 1000,
+		config = function()
+			require("catppuccin").setup {
+				flavor = "mocha",
+				integrations = {
+					cmp = true,
+					gitsigns = true,
+					nvimtree = true,
+					telescope = true,
+				}
+			}
+		end,
+	},
+}
