+++ dot_config/nvim/lua/plugins/git.lua
@@ -1,0 +1,18 @@
+return {
+  "tpope/vim-fugitive",
+  "tpope/vim-rhubarb",
+  {
+    "lewis6991/gitsigns.nvim",
+    config = function()
+      require('gitsigns').setup {
+        signs = {
+          add = { text = '+' },
+          change = { text = '~' },
+          delete = { text = '_' },
+          topdelete = { text = '‾' },
+          changedelete = { text = '~' },
+        },
+      }
+    end,
+  },
+}
