+++ dot_config/nvim/lua/plugins/init.lua
@@ -1,0 +1,3 @@
+return {
+	"folke/lazy.nvim"
+}
