+++ dot_config/nvim/lua/plugins/tmux.lua
@@ -1,0 +1,6 @@
+return {
+  "aserowy/tmux.nvim",
+  config = function()
+		require("tmux").setup()
+	end,
+}
